Make a working step towards incorporating twemoji emojis
Currently it's just the Mac-based implementation, but able to toggle
Twemoji use by uncommenting certain things. Tests will fail with Twemoji
As of now.
