+++ .gitignore
@@ -1,4 +1,4 @@
 dist/
 node_modules/
-
+assets/twemoji/*
 .DS_Store
