+++ package.json
@@ -31,7 +31,7 @@
         "app-root-path": "^3.1.0",
         "babel-jest": "^29.7.0",
         "babel-loader": "^9.1.3",
-        "chromedriver": "^125.0.0",
+        "chromedriver": "^128.0.1",
         "copy-webpack-plugin": "^11.0.0",
         "css-loader": "^6.8.1",
         "eslint": "^8.56.0",
