+++ src/contentScript/components/EmojiPicker/Emoji.module.css
@@ -1,0 +1,30 @@
+.emojiItem {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+
+    transition: background-color 0.3s ease;
+    text-align: center;
+    width: 32px;
+    height: 32px;
+    font-family: "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji", sans-serif !important;
+}
+
+.emojiWrapper {
+    font-size: 24px;
+    line-height: 24px;
+    vertical-align: middle;
+    display: inline-block;
+    width: 24px;
+    height: 24px;
+    text-align: center;
+}
+
+.emojiWrapper {
+    cursor: pointer;
+}
+
+.emojiItem:hover {
+    background-color: rgba(169, 169, 255, 0.829);
+    border-radius: 6px;
+}
