+++ src/contentScript/components/EmojiPicker/EmojiPicker.js
@@ -5,10 +5,10 @@ import { EMOJI_PICKER_ID, EMOJI_REMOVE_BUTTON_ID, SEARCH_BAR_ID, SEARCH_RESULTS_
 import classNames from 'classnames';
 import { getLogger } from '../../../log';
 import { findMatchingEmojis } from '../../emojiSearch';
-import twemoji from 'twemoji';
+import { Emoji } from './Emoji';
 
 // eslint-disable-next-line no-unused-vars
-const log = getLogger('EmojiPicker', 'debug');
+export const log = getLogger('EmojiPicker', 'debug');
 
 const EmojiPicker = ({ onEmojiClick, onRemoveEmoji }) => {
     const [searchValue, setSearchValue] = useState('');
@@ -112,57 +112,6 @@ const SearchBar = ({ onSearchBarChanged }) => {
 
 SearchBar.propTypes = {
     onSearchBarChanged: PropTypes.func.isRequired,
-}
-
-const Emoji = ({ emoji, onClick }) => {
-    const strippedShortcode = emoji.shortcode.substring(1, emoji.shortcode.length - 1);
-    const twemojiImg = twemoji.parse(emoji.emoji);
-
-    log.debug('Twemoji emoji:', twemojiImg, typeof twemojiImg);
-
-    const parseTwemojiOutput = (twemojiOutput) => {
-        const parser = new DOMParser();
-        const parsedDoc = parser.parseFromString(twemojiOutput, 'text/html');
-        log.debug('Parsed document:', parsedDoc);
-
-        const imgElement = parsedDoc.querySelector('img');
-        log.debug('Image element:', imgElement);
-
-
-        if (imgElement) {
-            log.debug('Image src:', imgElement.src);
-            log.debug('Image alt:', imgElement.alt);
-            return <img className={styles.emoji} draggable={false} src={imgElement.src} alt={imgElement.alt} />;
-        } else {
-            log.debug('ImgElement is null:', imgElement);
-            return null;
-        }
-
-    };
-    
-    const parseOutput = parseTwemojiOutput(twemojiImg);
-    log.debug('Parsed twemoji:', parseOutput);
-
-    return (
-        <div 
-            className={styles.emojiItem}
-            data-unicode={emoji.unicode} 
-            data-shortcode={emoji.shortcode}
-            onClick={() => onClick(emoji.emoji)}
-            title={strippedShortcode}
-        >
-            <span className={styles.emojiWrapper}>
-                {parseOutput}
-            {/* <span className={styles.emojiWrapper} dangerouslySetInnerHTML={{ __html: twemojiImg }}> */}
-                {/* {emoji.emoji} */}
-            </span>
-        </div>
-    );
-}
-
-Emoji.propTypes = {
-    emoji: PropTypes.object.isRequired,
-    onClick: PropTypes.func.isRequired,
 }
 
 export default EmojiPicker;
